implement adding new payment record
