--- app/lib/database.js
@@ -1,1659 +1,0 @@
-const users = [
-  {
-    id: "1a2b3c4d",
-    firstname: "Dwayne",
-    lastname: "Heibroch",
-    email: "dwayne@payproff.com",
-    password: "password",
-    imageUrl: "/dwayne.png",
-  },
-  {
-    id: "5e6f7g8h",
-    firstname: "John",
-    lastname: "Smith",
-    email: "john@payproff.com",
-    password: "password",
-    imageUrl: "/johnsmith.jpg",
-  },
-];
-
-const customers = [
-  {
-    id: "1a2b3c4d",
-    firstname: "John",
-    lastname: "Doe",
-    email: "john.doe@example.com",
-    imageUrl: "https://example.com/images/johndoe.jpg",
-  },
-  {
-    id: "5e6f7g8h",
-    firstname: "Jane",
-    lastname: "Smith",
-    email: "jane.smith@example.com",
-    imageUrl: "https://example.com/images/janesmith.jpg",
-  },
-  {
-    id: "9i0j1k2l",
-    firstname: "Michael",
-    lastname: "Johnson",
-    email: "michael.johnson@example.com",
-    imageUrl: "https://example.com/images/michaeljohnson.jpg",
-  },
-  {
-    id: "3m4n5o6p",
-    firstname: "Emily",
-    lastname: "Davis",
-    email: "emily.davis@example.com",
-    imageUrl: "https://example.com/images/emilydavis.jpg",
-  },
-  {
-    id: "7q8r9s0t",
-    firstname: "Chris",
-    lastname: "Brown",
-    email: "chris.brown@example.com",
-    imageUrl: "https://example.com/images/chrisbrown.jpg",
-  },
-  {
-    id: "1u2v3w4x",
-    firstname: "Sarah",
-    lastname: "Wilson",
-    email: "sarah.wilson@example.com",
-    imageUrl: "https://example.com/images/sarahwilson.jpg",
-  },
-  {
-    id: "5y6z7a8b",
-    firstname: "David",
-    lastname: "Martinez",
-    email: "david.martinez@example.com",
-    imageUrl: "https://example.com/images/davidmartinez.jpg",
-  },
-  {
-    id: "9c0d1e2f",
-    firstname: "Laura",
-    lastname: "Garcia",
-    email: "laura.garcia@example.com",
-    imageUrl: "https://example.com/images/lauragarcia.jpg",
-  },
-  {
-    id: "3g4h5i6j",
-    firstname: "James",
-    lastname: "Rodriguez",
-    email: "james.rodriguez@example.com",
-    imageUrl: "https://example.com/images/jamesrodriguez.jpg",
-  },
-  {
-    id: "7k8l9m0n",
-    firstname: "Olivia",
-    lastname: "Hernandez",
-    email: "olivia.hernandez@example.com",
-    imageUrl: "https://example.com/images/oliviahernandez.jpg",
-  },
-  {
-    id: "1p2q3r4s",
-    firstname: "William",
-    lastname: "Clark",
-    email: "william.clark@example.com",
-    imageUrl: "https://example.com/images/williamclark.jpg",
-  },
-  {
-    id: "5t6u7v8w",
-    firstname: "Sophia",
-    lastname: "Lewis",
-    email: "sophia.lewis@example.com",
-    imageUrl: "https://example.com/images/sophialewis.jpg",
-  },
-  {
-    id: "9x0y1z2a",
-    firstname: "Benjamin",
-    lastname: "Walker",
-    email: "benjamin.walker@example.com",
-    imageUrl: "https://example.com/images/benjaminwalker.jpg",
-  },
-  {
-    id: "3b4c5d6e",
-    firstname: "Ava",
-    lastname: "Hall",
-    email: "ava.hall@example.com",
-    imageUrl: "https://example.com/images/avahall.jpg",
-  },
-  {
-    id: "7f8g9h0i",
-    firstname: "Liam",
-    lastname: "Allen",
-    email: "liam.allen@example.com",
-    imageUrl: "https://example.com/images/liamallen.jpg",
-  },
-  {
-    id: "1j2k3l4m",
-    firstname: "Mia",
-    lastname: "Young",
-    email: "mia.young@example.com",
-    imageUrl: "https://example.com/images/miayoung.jpg",
-  },
-  {
-    id: "5n6o7p8q",
-    firstname: "Noah",
-    lastname: "King",
-    email: "noah.king@example.com",
-    imageUrl: "https://example.com/images/noahking.jpg",
-  },
-  {
-    id: "9r0s1t2u",
-    firstname: "Isabella",
-    lastname: "Scott",
-    email: "isabella.scott@example.com",
-    imageUrl: "https://example.com/images/isabellascott.jpg",
-  },
-  {
-    id: "3v4w5x6y",
-    firstname: "Lucas",
-    lastname: "Green",
-    email: "lucas.green@example.com",
-    imageUrl: "https://example.com/images/lucasgreen.jpg",
-  },
-  {
-    id: "7z8a9b0c",
-    firstname: "Amelia",
-    lastname: "Adams",
-    email: "amelia.adams@example.com",
-    imageUrl: "https://example.com/images/ameliaadams.jpg",
-  },
-  {
-    id: "1d2e3f4g",
-    firstname: "Mason",
-    lastname: "Baker",
-    email: "mason.baker@example.com",
-    imageUrl: "https://example.com/images/masonbaker.jpg",
-  },
-  {
-    id: "5h6i7j8k",
-    firstname: "Harper",
-    lastname: "Nelson",
-    email: "harper.nelson@example.com",
-    imageUrl: "https://example.com/images/harpernelson.jpg",
-  },
-  {
-    id: "9l0m1n2o",
-    firstname: "Ethan",
-    lastname: "Carter",
-    email: "ethan.carter@example.com",
-    imageUrl: "https://example.com/images/ethancarter.jpg",
-  },
-  {
-    id: "3p4q5r6s",
-    firstname: "Evelyn",
-    lastname: "Mitchell",
-    email: "evelyn.mitchell@example.com",
-    imageUrl: "https://example.com/images/evelynmitchell.jpg",
-  },
-  {
-    id: "7t8u9v0w",
-    firstname: "Logan",
-    lastname: "Perez",
-    email: "logan.perez@example.com",
-    imageUrl: "https://example.com/images/loganperez.jpg",
-  },
-  {
-    id: "1x2y3z4a",
-    firstname: "Abigail",
-    lastname: "Roberts",
-    email: "abigail.roberts@example.com",
-    imageUrl: "https://example.com/images/abigailroberts.jpg",
-  },
-  {
-    id: "5b6c7d8e",
-    firstname: "Alexander",
-    lastname: "Turner",
-    email: "alexander.turner@example.com",
-    imageUrl: "https://example.com/images/alexanderturner.jpg",
-  },
-  {
-    id: "9f0g1h2i",
-    firstname: "Charlotte",
-    lastname: "Phillips",
-    email: "charlotte.phillips@example.com",
-    imageUrl: "https://example.com/images/charlottephillips.jpg",
-  },
-  {
-    id: "3j4k5l6m",
-    firstname: "Daniel",
-    lastname: "Campbell",
-    email: "daniel.campbell@example.com",
-    imageUrl: "https://example.com/images/danielcampbell.jpg",
-  },
-  {
-    id: "7n8o9p0q",
-    firstname: "Sofia",
-    lastname: "Parker",
-    email: "sofia.parker@example.com",
-    imageUrl: "https://example.com/images/sofiaparker.jpg",
-  },
-  {
-    id: "1r2s3t4u",
-    firstname: "Henry",
-    lastname: "Evans",
-    email: "henry.evans@example.com",
-    imageUrl: "https://example.com/images/henryevans.jpg",
-  },
-  {
-    id: "5v6w7x8y",
-    firstname: "Avery",
-    lastname: "Edwards",
-    email: "avery.edwards@example.com",
-    imageUrl: "https://example.com/images/averyedwards.jpg",
-  },
-  {
-    id: "9z0a1b2c",
-    firstname: "Sebastian",
-    lastname: "Collins",
-    email: "sebastian.collins@example.com",
-    imageUrl: "https://example.com/images/sebastiancollins.jpg",
-  },
-  {
-    id: "3d4e5f6g",
-    firstname: "Ella",
-    lastname: "Stewart",
-    email: "ella.stewart@example.com",
-    imageUrl: "https://example.com/images/ellastewart.jpg",
-  },
-  {
-    id: "7h8i9j0k",
-    firstname: "Jack",
-    lastname: "Sanchez",
-    email: "jack.sanchez@example.com",
-    imageUrl: "https://example.com/images/jacksanchez.jpg",
-  },
-  {
-    id: "1l2m3n4o",
-    firstname: "Grace",
-    lastname: "Morris",
-    email: "grace.morris@example.com",
-    imageUrl: "https://example.com/images/gracemorris.jpg",
-  },
-  {
-    id: "5p6q7r8s",
-    firstname: "Oliver",
-    lastname: "Murphy",
-    email: "oliver.murphy@example.com",
-    imageUrl: "https://example.com/images/olivermurphy.jpg",
-  },
-  {
-    id: "9t0u1v2w",
-    firstname: "Scarlett",
-    lastname: "Cook",
-    email: "scarlett.cook@example.com",
-    imageUrl: "https://example.com/images/scarlettcook.jpg",
-  },
-  {
-    id: "3x4y5z6a",
-    firstname: "Elijah",
-    lastname: "Rogers",
-    email: "elijah.rogers@example.com",
-    imageUrl: "https://example.com/images/elijahrogers.jpg",
-  },
-  {
-    id: "7b8c9d0e",
-    firstname: "Chloe",
-    lastname: "Reed",
-    email: "chloe.reed@example.com",
-    imageUrl: "https://example.com/images/chloereed.jpg",
-  },
-  {
-    id: "1f2g3h4i",
-    firstname: "Jacob",
-    lastname: "Morgan",
-    email: "jacob.morgan@example.com",
-    imageUrl: "https://example.com/images/jacobmorgan.jpg",
-  },
-  {
-    id: "5j6k7l8m",
-    firstname: "Lily",
-    lastname: "Bell",
-    email: "lily.bell@example.com",
-    imageUrl: "https://example.com/images/lilybell.jpg",
-  },
-  {
-    id: "9n0o1p2q",
-    firstname: "Samuel",
-    lastname: "Bailey",
-    email: "samuel.bailey@example.com",
-    imageUrl: "https://example.com/images/samuelbailey.jpg",
-  },
-  {
-    id: "3r4s5t6u",
-    firstname: "Hannah",
-    lastname: "Rivera",
-    email: "hannah.rivera@example.com",
-    imageUrl: "https://example.com/images/hannahrivera.jpg",
-  },
-  {
-    id: "7v8w9x0y",
-    firstname: "Matthew",
-    lastname: "Cooper",
-    email: "matthew.cooper@example.com",
-    imageUrl: "https://example.com/images/matthewcooper.jpg",
-  },
-  {
-    id: "1z2a3b4c",
-    firstname: "Aria",
-    lastname: "Richardson",
-    email: "aria.richardson@example.com",
-    imageUrl: "https://example.com/images/ariarichardson.jpg",
-  },
-  {
-    id: "5d6e7f8g",
-    firstname: "Anthony",
-    lastname: "Cox",
-    email: "anthony.cox@example.com",
-    imageUrl: "https://example.com/images/anthonycox.jpg",
-  },
-  {
-    id: "9h0i1j2k",
-    firstname: "Ella",
-    lastname: "Howard",
-    email: "ella.howard@example.com",
-    imageUrl: "https://example.com/images/ellahoward.jpg",
-  },
-  {
-    id: "3l4m5n6o",
-    firstname: "Ryan",
-    lastname: "Ward",
-    email: "ryan.ward@example.com",
-    imageUrl: "https://example.com/images/ryanward.jpg",
-  },
-  {
-    id: "7p8q9r0s",
-    firstname: "Zoe",
-    lastname: "Torres",
-    email: "zoe.torres@example.com",
-    imageUrl: "https://example.com/images/zoetorres.jpg",
-  },
-  {
-    id: "1t2u3v4w",
-    firstname: "Andrew",
-    lastname: "Peterson",
-    email: "andrew.peterson@example.com",
-    imageUrl: "https://example.com/images/andrewpeterson.jpg",
-  },
-  {
-    id: "5x6y7z8a",
-    firstname: "Madison",
-    lastname: "Gray",
-    email: "madison.gray@example.com",
-    imageUrl: "https://example.com/images/madisongray.jpg",
-  },
-];
-
-const payments = [
-  {
-    id: "o2nLvC6R",
-    buyerFirstname: "Jessica",
-    buyerLastname: "Martin",
-    sellerFirstname: "John",
-    sellerLastname: "Lee",
-    buyerEmail: "jessica.martin@example.com",
-    sellerEmail: "john.lee@example.com",
-    date: "2024-05-18",
-    amount: 317,
-    trustAccount: "Stripe",
-    status: "paid",
-  },
-  {
-    id: "DzqfFtR7",
-    buyerFirstname: "Jennifer",
-    buyerLastname: "Hill",
-    sellerFirstname: "David",
-    sellerLastname: "Anderson",
-    buyerEmail: "jennifer.hill@example.com",
-    sellerEmail: "david.anderson@example.com",
-    date: "2024-05-04",
-    amount: 461,
-    trustAccount: "Venmo",
-    status: "pending",
-  },
-  {
-    id: "oY6M1vHk",
-    buyerFirstname: "John",
-    buyerLastname: "Wilson",
-    sellerFirstname: "Kimberly",
-    sellerLastname: "Gonzalez",
-    buyerEmail: "john.wilson@example.com",
-    sellerEmail: "kimberly.gonzalez@example.com",
-    date: "2024-05-03",
-    amount: 427,
-    trustAccount: "Stripe",
-    status: "disputed",
-  },
-  {
-    id: "TlS5y1oR",
-    buyerFirstname: "Nicole",
-    buyerLastname: "Johnson",
-    sellerFirstname: "Timothy",
-    sellerLastname: "Garcia",
-    buyerEmail: "nicole.johnson@example.com",
-    sellerEmail: "timothy.garcia@example.com",
-    date: "2024-05-08",
-    amount: 408,
-    trustAccount: "PayPal",
-    status: "paid",
-  },
-  {
-    id: "R82VgeMh",
-    buyerFirstname: "William",
-    buyerLastname: "Walker",
-    sellerFirstname: "Jennifer",
-    sellerLastname: "Wilson",
-    buyerEmail: "william.walker@example.com",
-    sellerEmail: "jennifer.wilson@example.com",
-    date: "2024-05-02",
-    amount: 212,
-    trustAccount: "Venmo",
-    status: "pending",
-  },
-  {
-    id: "o3uEgpS4",
-    buyerFirstname: "Michael",
-    buyerLastname: "Lewis",
-    sellerFirstname: "Amanda",
-    sellerLastname: "Martinez",
-    buyerEmail: "michael.lewis@example.com",
-    sellerEmail: "amanda.martinez@example.com",
-    date: "2024-05-07",
-    amount: 157,
-    trustAccount: "Stripe",
-    status: "disputed",
-  },
-  {
-    id: "JYzLW8Go",
-    buyerFirstname: "Jessica",
-    buyerLastname: "Young",
-    sellerFirstname: "William",
-    sellerLastname: "Hernandez",
-    buyerEmail: "jessica.young@example.com",
-    sellerEmail: "william.hernandez@example.com",
-    date: "2024-05-23",
-    amount: 180,
-    trustAccount: "PayPal",
-    status: "paid",
-  },
-  {
-    id: "tU7VsEjN",
-    buyerFirstname: "Jennifer",
-    buyerLastname: "Jackson",
-    sellerFirstname: "David",
-    sellerLastname: "Garcia",
-    buyerEmail: "jennifer.jackson@example.com",
-    sellerEmail: "david.garcia@example.com",
-    date: "2024-05-10",
-    amount: 158,
-    trustAccount: "Stripe",
-    status: "pending",
-  },
-  {
-    id: "m8L5gDzF",
-    buyerFirstname: "Christopher",
-    buyerLastname: "Brown",
-    sellerFirstname: "Mary",
-    sellerLastname: "Hernandez",
-    buyerEmail: "christopher.brown@example.com",
-    sellerEmail: "mary.hernandez@example.com",
-    date: "2024-05-25",
-    amount: 438,
-    trustAccount: "Venmo",
-    status: "disputed",
-  },
-  {
-    id: "U9bE4sNc",
-    buyerFirstname: "Emily",
-    buyerLastname: "Wilson",
-    sellerFirstname: "David",
-    sellerLastname: "White",
-    buyerEmail: "emily.wilson@example.com",
-    sellerEmail: "david.white@example.com",
-    date: "2024-05-15",
-    amount: 188,
-    trustAccount: "PayPal",
-    status: "paid",
-  },
-  {
-    id: "FhI6pZlW",
-    buyerFirstname: "Mark",
-    buyerLastname: "Brown",
-    sellerFirstname: "Jennifer",
-    sellerLastname: "Clark",
-    buyerEmail: "mark.brown@example.com",
-    sellerEmail: "jennifer.clark@example.com",
-    date: "2024-05-12",
-    amount: 364,
-    trustAccount: "Stripe",
-    status: "pending",
-  },
-  {
-    id: "w4bYzU5J",
-    buyerFirstname: "Kimberly",
-    buyerLastname: "Young",
-    sellerFirstname: "Andrew",
-    sellerLastname: "Martinez",
-    buyerEmail: "kimberly.young@example.com",
-    sellerEmail: "andrew.martinez@example.com",
-    date: "2024-05-13",
-    amount: 389,
-    trustAccount: "Venmo",
-    status: "disputed",
-  },
-  {
-    id: "R6tS9wUy",
-    buyerFirstname: "John",
-    buyerLastname: "Gonzalez",
-    sellerFirstname: "Jennifer",
-    sellerLastname: "Martinez",
-    buyerEmail: "john.gonzalez@example.com",
-    sellerEmail: "jennifer.martinez@example.com",
-    date: "2024-05-21",
-    amount: 477,
-    trustAccount: "Stripe",
-    status: "paid",
-  },
-  {
-    id: "t1zD4rFm",
-    buyerFirstname: "William",
-    buyerLastname: "Wilson",
-    sellerFirstname: "Mary",
-    sellerLastname: "Johnson",
-    buyerEmail: "william.wilson@example.com",
-    sellerEmail: "mary.johnson@example.com",
-    date: "2024-05-09",
-    amount: 192,
-    trustAccount: "PayPal",
-    status: "pending",
-  },
-  {
-    id: "K8nJ3sYv",
-    buyerFirstname: "David",
-    buyerLastname: "Williams",
-    sellerFirstname: "Emma",
-    sellerLastname: "Brown",
-    buyerEmail: "david.williams@example.com",
-    sellerEmail: "emma.brown@example.com",
-    date: "2024-05-11",
-    amount: 254,
-    trustAccount: "Stripe",
-    status: "paid",
-  },
-  {
-    id: "S6gP2uAq",
-    buyerFirstname: "Christopher",
-    buyerLastname: "Taylor",
-    sellerFirstname: "Jennifer",
-    sellerLastname: "Martin",
-    buyerEmail: "christopher.taylor@example.com",
-    sellerEmail: "jennifer.martin@example.com",
-    date: "2024-05-20",
-    amount: 176,
-    trustAccount: "Venmo",
-    status: "pending",
-  },
-  {
-    id: "E9rG5sPw",
-    buyerFirstname: "Sarah",
-    buyerLastname: "Thomas",
-    sellerFirstname: "Daniel",
-    sellerLastname: "Lewis",
-    buyerEmail: "sarah.thomas@example.com",
-    sellerEmail: "daniel.lewis@example.com",
-    date: "2024-05-05",
-    amount: 289,
-    trustAccount: "PayPal",
-    status: "disputed",
-  },
-  {
-    id: "L7bT9mZx",
-    buyerFirstname: "Kimberly",
-    buyerLastname: "Hill",
-    sellerFirstname: "John",
-    sellerLastname: "Clark",
-    buyerEmail: "kimberly.hill@example.com",
-    sellerEmail: "john.clark@example.com",
-    date: "2024-05-24",
-    amount: 326,
-    trustAccount: "Stripe",
-    status: "paid",
-  },
-  {
-    id: "Q4hE7jRw",
-    buyerFirstname: "Matthew",
-    buyerLastname: "Garcia",
-    sellerFirstname: "Jennifer",
-    sellerLastname: "Hernandez",
-    buyerEmail: "matthew.garcia@example.com",
-    sellerEmail: "jennifer.hernandez@example.com",
-    date: "2024-05-07",
-    amount: 412,
-    trustAccount: "Venmo",
-    status: "pending",
-  },
-  {
-    id: "D2sH8gWj",
-    buyerFirstname: "Amanda",
-    buyerLastname: "White",
-    sellerFirstname: "William",
-    sellerLastname: "Moore",
-    buyerEmail: "amanda.white@example.com",
-    sellerEmail: "william.moore@example.com",
-    date: "2024-05-01",
-    amount: 147,
-    trustAccount: "PayPal",
-    status: "disputed",
-  },
-  {
-    id: "K5fW9rTe",
-    buyerFirstname: "Jessica",
-    buyerLastname: "Hernandez",
-    sellerFirstname: "David",
-    sellerLastname: "Brown",
-    buyerEmail: "jessica.hernandez@example.com",
-    sellerEmail: "david.brown@example.com",
-    date: "2024-05-16",
-    amount: 383,
-    trustAccount: "Stripe",
-    status: "paid",
-  },
-  {
-    id: "R7jL5bTf",
-    buyerFirstname: "John",
-    buyerLastname: "Wilson",
-    sellerFirstname: "Jennifer",
-    sellerLastname: "Lewis",
-    buyerEmail: "john.wilson@example.com",
-    sellerEmail: "jennifer.lewis@example.com",
-    date: "2024-05-14",
-    amount: 199,
-    trustAccount: "Venmo",
-    status: "pending",
-  },
-  {
-    id: "M3hK8vLc",
-    buyerFirstname: "Emily",
-    buyerLastname: "Gonzalez",
-    sellerFirstname: "William",
-    sellerLastname: "White",
-    buyerEmail: "emily.gonzalez@example.com",
-    sellerEmail: "william.white@example.com",
-    date: "2024-05-03",
-    amount: 237,
-    trustAccount: "PayPal",
-    status: "disputed",
-  },
-  {
-    id: "m3n4o5p6",
-    buyerFirstname: "Diana",
-    buyerLastname: "Prince",
-    sellerFirstname: "Steve",
-    sellerLastname: "Trevor",
-    buyerEmail: "diana.prince@example.com",
-    sellerEmail: "steve.trevor@example.com",
-    date: "2024-01-15",
-    amount: 320,
-    trustAccount: "Stripe",
-    status: "returned",
-  },
-  {
-    id: "q7r8s9t0",
-    buyerFirstname: "Natasha",
-    buyerLastname: "Romanoff",
-    sellerFirstname: "Clint",
-    sellerLastname: "Barton",
-    buyerEmail: "natasha.romanoff@example.com",
-    sellerEmail: "clint.barton@example.com",
-    date: "2024-05-02",
-    amount: 270,
-    trustAccount: "PayPal",
-    status: "returned",
-  },
-  {
-    id: "u1v2w3x4",
-    buyerFirstname: "Peter",
-    buyerLastname: "Parker",
-    sellerFirstname: "Mary",
-    sellerLastname: "Jane",
-    buyerEmail: "peter.parker@example.com",
-    sellerEmail: "mary.jane@example.com",
-    date: "2024-04-14",
-    amount: 100,
-    trustAccount: "Stripe",
-    status: "returned",
-  },
-  {
-    id: "W6vH3jNu",
-    buyerFirstname: "Michael",
-    buyerLastname: "Hill",
-    sellerFirstname: "Mary",
-    sellerLastname: "Gonzalez",
-    buyerEmail: "michael.hill@example.com",
-    sellerEmail: "mary.gonzalez@example.com",
-    date: "2024-05-22",
-    amount: 145,
-    trustAccount: "Stripe",
-    status: "paid",
-  },
-  {
-    id: "T4sF7gZj",
-    buyerFirstname: "Jennifer",
-    buyerLastname: "Gonzalez",
-    sellerFirstname: "Christopher",
-    sellerLastname: "Williams",
-    buyerEmail: "jennifer.gonzalez@example.com",
-    sellerEmail: "christopher.williams@example.com",
-    date: "2024-05-08",
-    amount: 298,
-    trustAccount: "Venmo",
-    status: "pending",
-  },
-  {
-    id: "Q9tN5sFv",
-    buyerFirstname: "David",
-    buyerLastname: "Hernandez",
-    sellerFirstname: "Mary",
-    sellerLastname: "Martinez",
-    buyerEmail: "david.hernandez@example.com",
-    sellerEmail: "mary.martinez@example.com",
-    date: "2024-05-17",
-    amount: 325,
-    trustAccount: "PayPal",
-    status: "disputed",
-  },
-  {
-    id: "F8mD2bLq",
-    buyerFirstname: "Kimberly",
-    buyerLastname: "Brown",
-    sellerFirstname: "William",
-    sellerLastname: "Martinez",
-    buyerEmail: "kimberly.brown@example.com",
-    sellerEmail: "william.martinez@example.com",
-    date: "2024-05-06",
-    amount: 279,
-    trustAccount: "Stripe",
-    status: "paid",
-  },
-  {
-    id: "E2bK7rNl",
-    buyerFirstname: "John",
-    buyerLastname: "Wilson",
-    sellerFirstname: "Jessica",
-    sellerLastname: "Martinez",
-    buyerEmail: "john.wilson@example.com",
-    sellerEmail: "jessica.martinez@example.com",
-    date: "2024-05-25",
-    amount: 143,
-    trustAccount: "Venmo",
-    status: "pending",
-  },
-  {
-    id: "K2sT8mPb",
-    buyerFirstname: "John",
-    buyerLastname: "Brown",
-    sellerFirstname: "Emily",
-    sellerLastname: "Clark",
-    buyerEmail: "john.brown@example.com",
-    sellerEmail: "emily.clark@example.com",
-    date: "2024-05-12",
-    amount: 287,
-    trustAccount: "PayPal",
-    status: "disputed",
-  },
-  {
-    id: "V6gP3tJz",
-    buyerFirstname: "Matthew",
-    buyerLastname: "Lewis",
-    sellerFirstname: "Jennifer",
-    sellerLastname: "Williams",
-    buyerEmail: "matthew.lewis@example.com",
-    sellerEmail: "jennifer.williams@example.com",
-    date: "2024-05-20",
-    amount: 295,
-    trustAccount: "Stripe",
-    status: "paid",
-  },
-  {
-    id: "Q9sP8mTb",
-    buyerFirstname: "Jessica",
-    buyerLastname: "Gonzalez",
-    sellerFirstname: "David",
-    sellerLastname: "Hill",
-    buyerEmail: "jessica.gonzalez@example.com",
-    sellerEmail: "david.hill@example.com",
-    date: "2024-05-01",
-    amount: 259,
-    trustAccount: "Venmo",
-    status: "pending",
-  },
-  {
-    id: "R5vT2sPk",
-    buyerFirstname: "Michael",
-    buyerLastname: "Martinez",
-    sellerFirstname: "Sarah",
-    sellerLastname: "White",
-    buyerEmail: "michael.martinez@example.com",
-    sellerEmail: "sarah.white@example.com",
-    date: "2024-05-18",
-    amount: 223,
-    trustAccount: "PayPal",
-    status: "disputed",
-  },
-  {
-    id: "W2sP7tJm",
-    buyerFirstname: "David",
-    buyerLastname: "Gonzalez",
-    sellerFirstname: "Jennifer",
-    sellerLastname: "Taylor",
-    buyerEmail: "david.gonzalez@example.com",
-    sellerEmail: "jennifer.taylor@example.com",
-    date: "2024-05-24",
-    amount: 268,
-    trustAccount: "Stripe",
-    status: "paid",
-  },
-  {
-    id: "N9vT5sPd",
-    buyerFirstname: "Sarah",
-    buyerLastname: "Johnson",
-    sellerFirstname: "Matthew",
-    sellerLastname: "Wilson",
-    buyerEmail: "sarah.johnson@example.com",
-    sellerEmail: "matthew.wilson@example.com",
-    date: "2024-05-11",
-    amount: 207,
-    trustAccount: "Venmo",
-    status: "pending",
-  },
-  {
-    id: "L3gT7sPz",
-    buyerFirstname: "Jennifer",
-    buyerLastname: "Brown",
-    sellerFirstname: "David",
-    sellerLastname: "Hernandez",
-    buyerEmail: "jennifer.brown@example.com",
-    sellerEmail: "david.hernandez@example.com",
-    date: "2024-05-22",
-    amount: 242,
-    trustAccount: "PayPal",
-    status: "disputed",
-  },
-  {
-    id: "A5hT9sPc",
-    buyerFirstname: "Daniel",
-    buyerLastname: "Martinez",
-    sellerFirstname: "Emily",
-    sellerLastname: "Taylor",
-    buyerEmail: "daniel.martinez@example.com",
-    sellerEmail: "emily.taylor@example.com",
-    date: "2024-05-05",
-    amount: 281,
-    trustAccount: "Stripe",
-    status: "paid",
-  },
-  {
-    id: "K6fP4tJv",
-    buyerFirstname: "Jessica",
-    buyerLastname: "Lewis",
-    sellerFirstname: "Michael",
-    sellerLastname: "Brown",
-    buyerEmail: "jessica.lewis@example.com",
-    sellerEmail: "michael.brown@example.com",
-    date: "2024-05-13",
-    amount: 298,
-    trustAccount: "Venmo",
-    status: "pending",
-  },
-  {
-    id: "v7w8x9y0",
-    buyerFirstname: "John",
-    buyerLastname: "Doe",
-    sellerFirstname: "Jane",
-    sellerLastname: "Smith",
-    buyerEmail: "john.doe@example.com",
-    sellerEmail: "jane.smith@example.com",
-    date: "2024-05-12",
-    amount: 275,
-    trustAccount: "Stripe",
-    status: "returned",
-  },
-  {
-    id: "a1b2c3d4",
-    buyerFirstname: "Sarah",
-    buyerLastname: "Connor",
-    sellerFirstname: "Kyle",
-    sellerLastname: "Reese",
-    buyerEmail: "sarah.connor@example.com",
-    sellerEmail: "kyle.reese@example.com",
-    date: "2024-04-20",
-    amount: 150,
-    trustAccount: "PayPal",
-    status: "returned",
-  },
-  {
-    id: "e5f6g7h8",
-    buyerFirstname: "Tony",
-    buyerLastname: "Stark",
-    sellerFirstname: "Pepper",
-    sellerLastname: "Potts",
-    buyerEmail: "tony.stark@example.com",
-    sellerEmail: "pepper.potts@example.com",
-    date: "2024-03-18",
-    amount: 450,
-    trustAccount: "Stripe",
-    status: "returned",
-  },
-  {
-    id: "M8hT5sPz",
-    buyerFirstname: "John",
-    buyerLastname: "Clark",
-    sellerFirstname: "Sarah",
-    sellerLastname: "Hill",
-    buyerEmail: "john.clark@example.com",
-    sellerEmail: "sarah.hill@example.com",
-    date: "2024-05-06",
-    amount: 254,
-    trustAccount: "PayPal",
-    status: "disputed",
-  },
-  {
-    id: "B7gP3tJa",
-    buyerFirstname: "David",
-    buyerLastname: "Hernandez",
-    sellerFirstname: "Jennifer",
-    sellerLastname: "Clark",
-    buyerEmail: "david.hernandez@example.com",
-    sellerEmail: "jennifer.clark@example.com",
-    date: "2024-05-25",
-    amount: 295,
-    trustAccount: "Stripe",
-    status: "paid",
-  },
-  {
-    id: "R9sP2mTz",
-    buyerFirstname: "Sarah",
-    buyerLastname: "Brown",
-    sellerFirstname: "Matthew",
-    sellerLastname: "Gonzalez",
-    buyerEmail: "sarah.brown@example.com",
-    sellerEmail: "matthew.gonzalez@example.com",
-    date: "2024-05-17",
-    amount: 279,
-    trustAccount: "Venmo",
-    status: "pending",
-  },
-  {
-    id: "W4vT8sPq",
-    buyerFirstname: "Daniel",
-    buyerLastname: "Lewis",
-    sellerFirstname: "Emily",
-    sellerLastname: "Hill",
-    buyerEmail: "daniel.lewis@example.com",
-    sellerEmail: "emily.hill@example.com",
-    date: "2024-05-04",
-    amount: 222,
-    trustAccount: "PayPal",
-    status: "disputed",
-  },
-  {
-    id: "T3gP7tJn",
-    buyerFirstname: "Jessica",
-    buyerLastname: "Wilson",
-    sellerFirstname: "David",
-    sellerLastname: "Clark",
-    buyerEmail: "jessica.wilson@example.com",
-    sellerEmail: "david.clark@example.com",
-    date: "2024-05-23",
-    amount: 249,
-    trustAccount: "Stripe",
-    status: "paid",
-  },
-  {
-    id: "Z8jT2qKm",
-    buyerFirstname: "David",
-    buyerLastname: "Jones",
-    sellerFirstname: "Jessica",
-    sellerLastname: "Williams",
-    buyerEmail: "david.jones@example.com",
-    sellerEmail: "jessica.williams@example.com",
-    date: "2024-05-20",
-    amount: 197,
-    trustAccount: "Stripe",
-    status: "paid",
-  },
-  {
-    id: "U4vR9jQm",
-    buyerFirstname: "Sarah",
-    buyerLastname: "Hernandez",
-    sellerFirstname: "Christopher",
-    sellerLastname: "White",
-    buyerEmail: "sarah.hernandez@example.com",
-    sellerEmail: "christopher.white@example.com",
-    date: "2024-05-19",
-    amount: 169,
-    trustAccount: "Venmo",
-    status: "pending",
-  },
-  {
-    id: "J3fV5mRq",
-    buyerFirstname: "Jennifer",
-    buyerLastname: "Lee",
-    sellerFirstname: "William",
-    sellerLastname: "Martinez",
-    buyerEmail: "jennifer.lee@example.com",
-    sellerEmail: "william.martinez@example.com",
-    date: "2024-05-16",
-    amount: 233,
-    trustAccount: "PayPal",
-    status: "disputed",
-  },
-  {
-    id: "N6sR2qVb",
-    buyerFirstname: "Daniel",
-    buyerLastname: "Hernandez",
-    sellerFirstname: "Mary",
-    sellerLastname: "Brown",
-    buyerEmail: "daniel.hernandez@example.com",
-    sellerEmail: "mary.brown@example.com",
-    date: "2024-05-03",
-    amount: 189,
-    trustAccount: "Stripe",
-    status: "paid",
-  },
-  {
-    id: "G9mR3jLn",
-    buyerFirstname: "Sarah",
-    buyerLastname: "Hill",
-    sellerFirstname: "John",
-    sellerLastname: "Wilson",
-    buyerEmail: "sarah.hill@example.com",
-    sellerEmail: "john.wilson@example.com",
-    date: "2024-05-02",
-    amount: 224,
-    trustAccount: "Venmo",
-    status: "pending",
-  },
-  {
-    id: "S4jN8sUq",
-    buyerFirstname: "David",
-    buyerLastname: "Martin",
-    sellerFirstname: "Jessica",
-    sellerLastname: "Clark",
-    buyerEmail: "david.martin@example.com",
-    sellerEmail: "jessica.clark@example.com",
-    date: "2024-05-01",
-    amount: 259,
-    trustAccount: "PayPal",
-    status: "disputed",
-  },
-  {
-    id: "F7vT3mRz",
-    buyerFirstname: "Jennifer",
-    buyerLastname: "Hernandez",
-    sellerFirstname: "David",
-    sellerLastname: "White",
-    buyerEmail: "jennifer.hernandez@example.com",
-    sellerEmail: "david.white@example.com",
-    date: "2024-05-23",
-    amount: 202,
-    trustAccount: "Stripe",
-    status: "paid",
-  },
-  {
-    id: "T6jL9vSk",
-    buyerFirstname: "Kimberly",
-    buyerLastname: "Garcia",
-    sellerFirstname: "Christopher",
-    sellerLastname: "Lee",
-    buyerEmail: "kimberly.garcia@example.com",
-    sellerEmail: "christopher.lee@example.com",
-    date: "2024-05-10",
-    amount: 191,
-    trustAccount: "Venmo",
-    status: "pending",
-  },
-  {
-    id: "W5sR7tPq",
-    buyerFirstname: "Matthew",
-    buyerLastname: "Wilson",
-    sellerFirstname: "Jennifer",
-    sellerLastname: "Brown",
-    buyerEmail: "matthew.wilson@example.com",
-    sellerEmail: "jennifer.brown@example.com",
-    date: "2024-05-07",
-    amount: 236,
-    trustAccount: "PayPal",
-    status: "disputed",
-  },
-  {
-    id: "J9gM2rUc",
-    buyerFirstname: "Jessica",
-    buyerLastname: "Clark",
-    sellerFirstname: "David",
-    sellerLastname: "Taylor",
-    buyerEmail: "jessica.clark@example.com",
-    sellerEmail: "david.taylor@example.com",
-    date: "2024-05-15",
-    amount: 175,
-    trustAccount: "Stripe",
-    status: "paid",
-  },
-  {
-    id: "K8sU2jNz",
-    buyerFirstname: "Michael",
-    buyerLastname: "Martinez",
-    sellerFirstname: "Jennifer",
-    sellerLastname: "White",
-    buyerEmail: "michael.martinez@example.com",
-    sellerEmail: "jennifer.white@example.com",
-    date: "2024-05-24",
-    amount: 283,
-    trustAccount: "Venmo",
-    status: "pending",
-  },
-  {
-    id: "E6vT5sPb",
-    buyerFirstname: "David",
-    buyerLastname: "Brown",
-    sellerFirstname: "Kimberly",
-    sellerLastname: "Hernandez",
-    buyerEmail: "david.brown@example.com",
-    sellerEmail: "kimberly.hernandez@example.com",
-    date: "2024-05-06",
-    amount: 215,
-    trustAccount: "PayPal",
-    status: "disputed",
-  },
-  {
-    id: "Q5rN8tJm",
-    buyerFirstname: "John",
-    buyerLastname: "Clark",
-    sellerFirstname: "Jessica",
-    sellerLastname: "Gonzalez",
-    buyerEmail: "john.clark@example.com",
-    sellerEmail: "jessica.gonzalez@example.com",
-    date: "2024-05-05",
-    amount: 248,
-    trustAccount: "Stripe",
-    status: "paid",
-  },
-  {
-    id: "P3jM7rTs",
-    buyerFirstname: "Jennifer",
-    buyerLastname: "Martin",
-    sellerFirstname: "William",
-    sellerLastname: "Johnson",
-    buyerEmail: "jennifer.martin@example.com",
-    sellerEmail: "william.johnson@example.com",
-    date: "2024-05-04",
-    amount: 273,
-    trustAccount: "Venmo",
-    status: "pending",
-  },
-  {
-    id: "H4gT9sPl",
-    buyerFirstname: "Michael",
-    buyerLastname: "Brown",
-    sellerFirstname: "Jessica",
-    sellerLastname: "Lewis",
-    buyerEmail: "michael.brown@example.com",
-    sellerEmail: "jessica.lewis@example.com",
-    date: "2024-05-21",
-    amount: 289,
-    trustAccount: "Stripe",
-    status: "paid",
-  },
-  {
-    id: "N7vP5tJn",
-    buyerFirstname: "Daniel",
-    buyerLastname: "Hernandez",
-    sellerFirstname: "Sarah",
-    sellerLastname: "Garcia",
-    buyerEmail: "daniel.hernandez@example.com",
-    sellerEmail: "sarah.garcia@example.com",
-    date: "2024-05-14",
-    amount: 237,
-    trustAccount: "Venmo",
-    status: "pending",
-  },
-  {
-    id: "W5sT9mPv",
-    buyerFirstname: "Jessica",
-    buyerLastname: "Clark",
-    sellerFirstname: "David",
-    sellerLastname: "Martin",
-    buyerEmail: "jessica.clark@example.com",
-    sellerEmail: "david.martin@example.com",
-    date: "2024-05-23",
-    amount: 263,
-    trustAccount: "PayPal",
-    status: "disputed",
-  },
-  {
-    id: "M8hP3tJs",
-    buyerFirstname: "Sarah",
-    buyerLastname: "Lewis",
-    sellerFirstname: "Matthew",
-    sellerLastname: "White",
-    buyerEmail: "sarah.lewis@example.com",
-    sellerEmail: "matthew.white@example.com",
-    date: "2024-05-20",
-    amount: 276,
-    trustAccount: "Stripe",
-    status: "paid",
-  },
-  {
-    id: "L6gT9sPr",
-    buyerFirstname: "David",
-    buyerLastname: "Johnson",
-    sellerFirstname: "Jennifer",
-    sellerLastname: "Hernandez",
-    buyerEmail: "david.johnson@example.com",
-    sellerEmail: "jennifer.hernandez@example.com",
-    date: "2024-05-03",
-    amount: 215,
-    trustAccount: "Venmo",
-    status: "pending",
-  },
-  {
-    id: "H9sP4tJm",
-    buyerFirstname: "Michael",
-    buyerLastname: "Gonzalez",
-    sellerFirstname: "Sarah",
-    sellerLastname: "Brown",
-    buyerEmail: "michael.gonzalez@example.com",
-    sellerEmail: "sarah.brown@example.com",
-    date: "2024-05-22",
-    amount: 291,
-    trustAccount: "PayPal",
-    status: "disputed",
-  },
-  {
-    id: "F7vT8mPk",
-    buyerFirstname: "Daniel",
-    buyerLastname: "Taylor",
-    sellerFirstname: "Emily",
-    sellerLastname: "Clark",
-    buyerEmail: "daniel.taylor@example.com",
-    sellerEmail: "emily.clark@example.com",
-    date: "2024-05-01",
-    amount: 228,
-    trustAccount: "Stripe",
-    status: "paid",
-  },
-  {
-    id: "R3gP6tJz",
-    buyerFirstname: "Jessica",
-    buyerLastname: "Martinez",
-    sellerFirstname: "David",
-    sellerLastname: "Wilson",
-    buyerEmail: "jessica.martinez@example.com",
-    sellerEmail: "david.wilson@example.com",
-    date: "2024-05-09",
-    amount: 242,
-    trustAccount: "Venmo",
-    status: "pending",
-  },
-  {
-    id: "B5hT7sPq",
-    buyerFirstname: "Sarah",
-    buyerLastname: "Garcia",
-    sellerFirstname: "Michael",
-    sellerLastname: "Martinez",
-    buyerEmail: "sarah.garcia@example.com",
-    sellerEmail: "michael.martinez@example.com",
-    date: "2024-05-26",
-    amount: 263,
-    trustAccount: "PayPal",
-    status: "disputed",
-  },
-  {
-    id: "Q8sP4tJm",
-    buyerFirstname: "David",
-    buyerLastname: "Brown",
-    sellerFirstname: "Jessica",
-    sellerLastname: "Gonzalez",
-    buyerEmail: "david.brown@example.com",
-    sellerEmail: "jessica.gonzalez@example.com",
-    date: "2024-05-15",
-    amount: 227,
-    trustAccount: "Stripe",
-    status: "paid",
-  },
-  {
-    id: "N9vT2sPc",
-    buyerFirstname: "Sarah",
-    buyerLastname: "Johnson",
-    sellerFirstname: "Daniel",
-    sellerLastname: "Lewis",
-    buyerEmail: "sarah.johnson@example.com",
-    sellerEmail: "daniel.lewis@example.com",
-    date: "2024-05-02",
-    amount: 256,
-    trustAccount: "Venmo",
-    status: "pending",
-  },
-  {
-    id: "J4gT8sPz",
-    buyerFirstname: "Michael",
-    buyerLastname: "Wilson",
-    sellerFirstname: "Sarah",
-    sellerLastname: "Clark",
-    buyerEmail: "michael.wilson@example.com",
-    sellerEmail: "sarah.clark@example.com",
-    date: "2024-05-21",
-    amount: 282,
-    trustAccount: "PayPal",
-    status: "disputed",
-  },
-  {
-    id: "W6vT3sPb",
-    buyerFirstname: "David",
-    buyerLastname: "Gonzalez",
-    sellerFirstname: "Jessica",
-    sellerLastname: "Martin",
-    buyerEmail: "david.gonzalez@example.com",
-    sellerEmail: "jessica.martin@example.com",
-    date: "2024-05-08",
-    amount: 219,
-    trustAccount: "Stripe",
-    status: "paid",
-  },
-  {
-    id: "K2sP9tJc",
-    buyerFirstname: "Sarah",
-    buyerLastname: "Brown",
-    sellerFirstname: "Matthew",
-    sellerLastname: "Johnson",
-    buyerEmail: "sarah.brown@example.com",
-    sellerEmail: "matthew.johnson@example.com",
-    date: "2024-05-05",
-    amount: 245,
-    trustAccount: "Venmo",
-    status: "pending",
-  },
-  {
-    id: "o2nLvC6R",
-    buyerFirstname: "Jessica",
-    buyerLastname: "Martin",
-    sellerFirstname: "John",
-    sellerLastname: "Lee",
-    buyerEmail: "jessica.martin@example.com",
-    sellerEmail: "john.lee@example.com",
-    date: "2024-05-18",
-    amount: 317,
-    trustAccount: "Stripe",
-    status: "paid",
-  },
-  {
-    id: "a1b2c3d4",
-    buyerFirstname: "Alice",
-    buyerLastname: "Smith",
-    sellerFirstname: "Bob",
-    sellerLastname: "Johnson",
-    buyerEmail: "alice.smith@example.com",
-    sellerEmail: "bob.johnson@example.com",
-    date: "2024-04-22",
-    amount: 150,
-    trustAccount: "PayPal",
-    status: "paid",
-  },
-  {
-    id: "e5f6g7h8",
-    buyerFirstname: "Michael",
-    buyerLastname: "Brown",
-    sellerFirstname: "Eve",
-    sellerLastname: "Davis",
-    buyerEmail: "michael.brown@example.com",
-    sellerEmail: "eve.davis@example.com",
-    date: "2024-03-19",
-    amount: 200,
-    trustAccount: "Stripe",
-    status: "pending",
-  },
-  {
-    id: "i9j0k1l2",
-    buyerFirstname: "Chris",
-    buyerLastname: "Wilson",
-    sellerFirstname: "Dana",
-    sellerLastname: "Taylor",
-    buyerEmail: "chris.wilson@example.com",
-    sellerEmail: "dana.taylor@example.com",
-    date: "2024-02-14",
-    amount: 425,
-    trustAccount: "PayPal",
-    status: "paid",
-  },
-  {
-    id: "m3n4o5p6",
-    buyerFirstname: "Emma",
-    buyerLastname: "Moore",
-    sellerFirstname: "Frank",
-    sellerLastname: "Anderson",
-    buyerEmail: "emma.moore@example.com",
-    sellerEmail: "frank.anderson@example.com",
-    date: "2024-01-30",
-    amount: 320,
-    trustAccount: "Stripe",
-    status: "returned",
-  },
-  {
-    id: "q7r8s9t0",
-    buyerFirstname: "Liam",
-    buyerLastname: "Thomas",
-    sellerFirstname: "Grace",
-    sellerLastname: "Jackson",
-    buyerEmail: "liam.thomas@example.com",
-    sellerEmail: "grace.jackson@example.com",
-    date: "2024-05-03",
-    amount: 270,
-    trustAccount: "PayPal",
-    status: "paid",
-  },
-  {
-    id: "u1v2w3x4",
-    buyerFirstname: "Sophia",
-    buyerLastname: "White",
-    sellerFirstname: "Henry",
-    sellerLastname: "Harris",
-    buyerEmail: "sophia.white@example.com",
-    sellerEmail: "henry.harris@example.com",
-    date: "2024-04-15",
-    amount: 100,
-    trustAccount: "Stripe",
-    status: "pending",
-  },
-  {
-    id: "y5z6a7b8",
-    buyerFirstname: "James",
-    buyerLastname: "Martin",
-    sellerFirstname: "Ivy",
-    sellerLastname: "Thompson",
-    buyerEmail: "james.martin@example.com",
-    sellerEmail: "ivy.thompson@example.com",
-    date: "2024-03-07",
-    amount: 500,
-    trustAccount: "PayPal",
-    status: "paid",
-  },
-  {
-    id: "c9d0e1f2",
-    buyerFirstname: "Olivia",
-    buyerLastname: "Garcia",
-    sellerFirstname: "Jack",
-    sellerLastname: "Martinez",
-    buyerEmail: "olivia.garcia@example.com",
-    sellerEmail: "jack.martinez@example.com",
-    date: "2024-02-26",
-    amount: 350,
-    trustAccount: "Stripe",
-    status: "paid",
-  },
-  {
-    id: "g3h4i5j6",
-    buyerFirstname: "William",
-    buyerLastname: "Rodriguez",
-    sellerFirstname: "Karen",
-    sellerLastname: "Martinez",
-    buyerEmail: "william.rodriguez@example.com",
-    sellerEmail: "karen.martinez@example.com",
-    date: "2024-01-19",
-    amount: 275,
-    trustAccount: "PayPal",
-    status: "returned",
-  },
-  {
-    id: "k7l8m9n0",
-    buyerFirstname: "Isabella",
-    buyerLastname: "Martinez",
-    sellerFirstname: "Leo",
-    sellerLastname: "Hernandez",
-    buyerEmail: "isabella.martinez@example.com",
-    sellerEmail: "leo.hernandez@example.com",
-    date: "2024-05-25",
-    amount: 230,
-    trustAccount: "Stripe",
-    status: "paid",
-  },
-  {
-    id: "o1p2q3r4",
-    buyerFirstname: "Benjamin",
-    buyerLastname: "Lopez",
-    sellerFirstname: "Mia",
-    sellerLastname: "Gonzalez",
-    buyerEmail: "benjamin.lopez@example.com",
-    sellerEmail: "mia.gonzalez@example.com",
-    date: "2024-04-17",
-    amount: 190,
-    trustAccount: "PayPal",
-    status: "pending",
-  },
-  {
-    id: "s5t6u7v8",
-    buyerFirstname: "Ava",
-    buyerLastname: "Perez",
-    sellerFirstname: "Nathan",
-    sellerLastname: "Wilson",
-    buyerEmail: "ava.perez@example.com",
-    sellerEmail: "nathan.wilson@example.com",
-    date: "2024-03-25",
-    amount: 210,
-    trustAccount: "Stripe",
-    status: "paid",
-  },
-  {
-    id: "w9x0y1z2",
-    buyerFirstname: "Lucas",
-    buyerLastname: "Sanchez",
-    sellerFirstname: "Olivia",
-    sellerLastname: "White",
-    buyerEmail: "lucas.sanchez@example.com",
-    sellerEmail: "olivia.white@example.com",
-    date: "2024-02-03",
-    amount: 315,
-    trustAccount: "PayPal",
-    status: "returned",
-  },
-  {
-    id: "a3b4c5d6",
-    buyerFirstname: "Emily",
-    buyerLastname: "Clark",
-    sellerFirstname: "Paul",
-    sellerLastname: "Davis",
-    buyerEmail: "emily.clark@example.com",
-    sellerEmail: "paul.davis@example.com",
-    date: "2024-01-21",
-    amount: 140,
-    trustAccount: "Stripe",
-    status: "paid",
-  },
-  {
-    id: "e7f8g9h0",
-    buyerFirstname: "Henry",
-    buyerLastname: "Lewis",
-    sellerFirstname: "Sophia",
-    sellerLastname: "Walker",
-    buyerEmail: "henry.lewis@example.com",
-    sellerEmail: "sophia.walker@example.com",
-    date: "2024-05-10",
-    amount: 280,
-    trustAccount: "PayPal",
-    status: "pending",
-  },
-  {
-    id: "i1j2k3l4",
-    buyerFirstname: "Amelia",
-    buyerLastname: "Robinson",
-    sellerFirstname: "David",
-    sellerLastname: "Young",
-    buyerEmail: "amelia.robinson@example.com",
-    sellerEmail: "david.young@example.com",
-    date: "2024-04-05",
-    amount: 370,
-    trustAccount: "Stripe",
-    status: "paid",
-  },
-  {
-    id: "m5n6o7p8",
-    buyerFirstname: "Daniel",
-    buyerLastname: "Hall",
-    sellerFirstname: "Victoria",
-    sellerLastname: "Allen",
-    buyerEmail: "daniel.hall@example.com",
-    sellerEmail: "victoria.allen@example.com",
-    date: "2024-03-12",
-    amount: 255,
-    trustAccount: "PayPal",
-    status: "returned",
-  },
-  {
-    id: "q9r0s1t2",
-    buyerFirstname: "Mason",
-    buyerLastname: "King",
-    sellerFirstname: "Emily",
-    sellerLastname: "Scott",
-    buyerEmail: "mason.king@example.com",
-    sellerEmail: "emily.scott@example.com",
-    date: "2024-02-29",
-    amount: 295,
-    trustAccount: "Stripe",
-    status: "paid",
-  },
-  {
-    id: "u3v4w5x6",
-    buyerFirstname: "Elizabeth",
-    buyerLastname: "Green",
-    sellerFirstname: "James",
-    sellerLastname: "Adams",
-    buyerEmail: "elizabeth.green@example.com",
-    sellerEmail: "james.adams@example.com",
-    date: "2024-01-11",
-    amount: 310,
-    trustAccount: "PayPal",
-    status: "pending",
-  }
-
-];
-
-const revenue = [
-  { month: "Jan", revenue: 2000 },
-  { month: "Feb", revenue: 1800 },
-  { month: "Mar", revenue: 2200 },
-  { month: "Apr", revenue: 2500 },
-  { month: "May", revenue: 2300 },
-  { month: "Jun", revenue: 3200 },
-  { month: "Jul", revenue: 3500 },
-  { month: "Aug", revenue: 3700 },
-  { month: "Sep", revenue: 2500 },
-  { month: "Oct", revenue: 2800 },
-  { month: "Nov", revenue: 2800 },
-  { month: "Dec", revenue: 4800 },
-];
-
-module.exports = {
-  users,
-  customers,
-  payments,
-  revenue,
-};
