+++ package.json
@@ -16,7 +16,8 @@
     "react": "^18",
     "react-dom": "^18",
     "recharts": "^2.12.7",
-    "use-debounce": "^10.0.0"
+    "use-debounce": "^10.0.0",
+    "zod": "^3.23.8"
   },
   "devDependencies": {
     "@types/node": "^20",
